+++ ai_portfolio/src/app/globals.css
@@ -40,15 +40,6 @@ body,
   padding: 0;
 }
 
-h1 {
-  font-size: 8em;
-  width: 55rem;
-  margin: 0;
-  padding: 0;
-  color: white;
-  text-shadow: 1px 0px 35px #383838;
-}
-
 a {
   pointer-events: all;
   text-decoration: none;
