+++ ai_portfolio/src/app/layout.tsx
@@ -2,7 +2,6 @@ import type { Metadata } from 'next';
 import { Geist, Geist_Mono } from 'next/font/google';
 import Header from '@/components/layout/header/Header';
 import Footer from '@/components/layout/footer/Footer';
-import { ScrollVisibilityProvider } from '@/context/ScrollVisibilityContext';
 import './globals.css';
 
 const geistSans = Geist({
@@ -31,9 +30,7 @@ export default function RootLayout({
         className={`${geistSans.variable} ${geistMono.variable} antialiased`}
       >
         <Header />
-        <main>
-          <ScrollVisibilityProvider>{children}</ScrollVisibilityProvider>
-        </main>
+        <main>{children}</main>
         <Footer />
       </body>
     </html>
