+++ ai_portfolio/src/types/component.ts
@@ -1,0 +1,6 @@
+/**
+ * @param isVisible - id=helloが表示されているかどうか
+ */
+export interface TopSectionProps {
+  isVisible: boolean;
+}
